Tag resets and dependencies corrections.
