+++ src/System.UnitTests.Shared/Microsoft.Omex.System.UnitTests.Shared.csproj
@@ -8,14 +8,12 @@
   <ItemGroup Condition="'$(TargetFramework)' != '$(NetStandardVersion)'">
     <PackageReference Include="xunit.assert" Version="$(XunitAssertVersion)" />
     <PackageReference Include="xunit.core" Version="$(XunitCoreVersion)" />
-    <PackageReference Include="System.Collections.Immutable" Version="$(SystemCollectionsImmutableVersion)" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFramework)' == '$(NetStandardVersion)'">
     <PackageReference Include="xunit.assert" Version="$(XunitAssertVersionCore)" />
     <PackageReference Include="xunit.core" Version="$(XunitCoreVersionCore)" />
-    <PackageReference Include="System.Collections.Immutable" Version="$(SystemCollectionsImmutableVersionCore)" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\src\System\Microsoft.Omex.System.csproj" />
+    <ProjectReference Include="..\System\Microsoft.Omex.System.csproj" />
   </ItemGroup>
 </Project>
