+++ src/System/Microsoft.Omex.System.csproj
@@ -8,6 +8,6 @@
   <ItemGroup>
     <Content Include="Readme.txt" PackagePath="%(Identity)" />
     <PackageReference Include="Bond.Core.CSharp" Version="$(BondCoreCSharpVersion)" />
-    <PackageReference Include="System.Collections.Immutable" Version="$(SystemCollectionsImmutableVersionCore)" />
+    <PackageReference Include="System.Collections.Immutable" Version="$(SystemCollectionsImmutableVersion)" />
   </ItemGroup>
 </Project>
