+++ src/System/TimedScopes/TransactionData.cs
@@ -1,11 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-#region Using Directives
-
 using System;
-
-#endregion
 
 namespace Microsoft.Omex.System.TimedScopes
 {
